// submodules/common-dto/api-client/main/models/KeywordDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type KeywordDto = {
    /**
     * Id of keyword
     */
    id: number;
    /**
     * Localized title of keyword
     */
    title: string;
    /**
     * ID of image
     */
    image_id: string | null;
};